Revert "Add menu"

+++ src/entities/Entity.java
@@ -2,22 +2,22 @@ package entities;
 
 import java.awt.Color;
 import java.awt.Graphics;
+import java.awt.Rectangle;
 import java.awt.geom.Rectangle2D;
 
 public abstract class Entity {
-
-	protected float x, y;
+	
+	protected float x,y;
 	protected int width, height;
 	protected Rectangle2D.Float hitbox;
-
+	
 	public Entity(float x, float y, int width, int height) {
 		this.x = x;
 		this.y = y;
 		this.width = width;
 		this.height = height;
-
 	}
-
+	
 	protected void drawHitbox(Graphics g) {
 		// For debugging the hitbox
 		g.setColor(Color.PINK);
@@ -25,7 +25,7 @@ public abstract class Entity {
 
 	}
 
-	protected void initHitbox(float x, float y, int width, int height) {
+	protected void initHitbox(float x, float y, float width, float height) {
 		hitbox = new Rectangle2D.Float(x, y, width, height);
 	}
 
@@ -37,5 +37,4 @@ public abstract class Entity {
 	public Rectangle2D.Float getHitbox() {
 		return hitbox;
 	}
-
 }
